# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Site.Set Blog API",
    "description": "API para gerenciamento do blog do Site.Set",
    "version": "1.0.0",
    "contact": {
      "name": "Site.Set Team",
      "email": "dev@site.set",
      "url": "https://site.set.vercel.app"
    }
  },
  "servers": [
    {
      "url": "https://site-set.vercel.app/api",
      "description": "Production"
    },
    {
      "url": "http://localhost:3000/api",
      "description": "Development"
    }
  ],
  "tags": [
    {
      "name": "Posts",
      "description": "Endpoints para gerenciamento de posts do blog"
    },
    {
      "name": "Waitlist",
      "description": "Endpoints para gerenciamento da lista de espera"
    }
  ],
  "paths": {
    "/posts": {
      "get": {
        "operationId": "listPosts",
        "tags": ["Posts"],
        "summary": "Lista todos os posts",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "description": "Número máximo de posts a retornar",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de posts retornada com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Post"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Não autorizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "post": {
        "operationId": "createPost",
        "tags": ["Posts"],
        "summary": "Cria um novo post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePostInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Post criado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Post"
                }
              }
            }
          },
          "400": {
            "description": "Dados inválidos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationError"
                }
              }
            }
          },
          "401": {
            "description": "Não autorizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/posts/search": {
      "get": {
        "operationId": "searchPosts",
        "tags": ["Posts"],
        "summary": "Busca posts por termo",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "description": "Termo de busca",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de posts encontrados",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Post"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Não autorizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/posts/{slug}": {
      "get": {
        "operationId": "getPostBySlug",
        "tags": ["Posts"],
        "summary": "Busca um post pelo slug",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "description": "Slug do post",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Post encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Post"
                }
              }
            }
          },
          "404": {
            "description": "Post não encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Não autorizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "put": {
        "operationId": "updatePost",
        "tags": ["Posts"],
        "summary": "Atualiza um post",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "description": "Slug do post",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePostInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Post atualizado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Post"
                }
              }
            }
          },
          "404": {
            "description": "Post não encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Não autorizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "delete": {
        "operationId": "deletePost",
        "tags": ["Posts"],
        "summary": "Remove um post",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "description": "Slug do post",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Post removido com sucesso"
          },
          "404": {
            "description": "Post não encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Não autorizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/waitlist": {
      "get": {
        "operationId": "listWaitlistUsers",
        "tags": ["Waitlist"],
        "summary": "Lista todos os usuários da lista de espera",
        "responses": {
          "200": {
            "description": "Lista de usuários retornada com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WaitlistUser"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Não autorizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "post": {
        "operationId": "createWaitlistUser",
        "tags": ["Waitlist"],
        "summary": "Adiciona um novo usuário à lista de espera",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWaitlistUserInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Usuário adicionado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WaitlistUser"
                }
              }
            }
          },
          "200": {
            "description": "E-mail já cadastrado",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["message"],
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "E-mail já cadastrado"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "E-mail inválido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Post": {
        "type": "object",
        "required": ["id", "slug", "title", "content", "excerpt", "coverUrl", "createdAt", "updatedAt"],
        "properties": {
          "id": {
            "type": "string",
            "format": "cuid"
          },
          "slug": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "content": {
            "type": "string"
          },
          "excerpt": {
            "type": "string"
          },
          "coverUrl": {
            "type": "string",
            "format": "uri"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "CreatePostInput": {
        "type": "object",
        "required": ["slug", "title", "content", "excerpt", "coverUrl"],
        "properties": {
          "slug": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "content": {
            "type": "string"
          },
          "excerpt": {
            "type": "string"
          },
          "coverUrl": {
            "type": "string",
            "format": "uri"
          }
        }
      },
      "UpdatePostInput": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "content": {
            "type": "string"
          },
          "excerpt": {
            "type": "string"
          },
          "coverUrl": {
            "type": "string",
            "format": "uri"
          }
        }
      },
      "Error": {
        "type": "object",
        "required": ["error"],
        "properties": {
          "error": {
            "type": "string"
          }
        }
      },
      "ValidationError": {
        "type": "object",
        "required": ["error", "errors"],
        "properties": {
          "error": {
            "type": "string"
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "object"
            }
          }
        }
      },
      "WaitlistUser": {
        "type": "object",
        "required": ["id", "email", "createdAt"],
        "properties": {
          "id": {
            "type": "string",
            "format": "cuid"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "CreateWaitlistUserInput": {
        "type": "object",
        "required": ["email"],
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Use o token definido em API_AUTH_TOKEN nas variáveis de ambiente"
      }
    }
  }
}